Fixing a few things, adding basic features

+++ NForum.Core/Services/CategoryService.cs
@@ -15,49 +15,39 @@ namespace NForum.Core.Services {
 		protected readonly IUserProvider userProvider;
 		protected readonly IEventPublisher eventPublisher;
 		protected readonly ILogger logger;
+		protected readonly IPermissionService permService;
 
 		public CategoryService(IUserProvider userProvider,
 					ICategoryRepository categoryRepo,
 					IEventPublisher eventPublisher,
-					ILogger logger) {
+					ILogger logger,
+					IPermissionService permService) {
 
 			this.categoryRepo = categoryRepo;
 			this.userProvider = userProvider;
 			this.eventPublisher = eventPublisher;
 			this.logger = logger;
-		}
-
-		private Boolean IsSolutionAdministrator(User user) {
-			if (user == null) {
-				return false;
-			}
-			// TODO:
-			return true;
+			this.permService = permService;
 		}
 
 		/// <summary>
 		/// Method for creating a new category.
 		/// </summary>
-		/// <param name="board">The parent board of the category.</param>
 		/// <param name="name">The name of the category.</param>
 		/// <param name="description">The description of the category.</param>
 		/// <param name="sortOrder">The sort order/placement of the category.</param>
 		/// <returns>The newly created category.</returns>
-		public Category Create(Board board, String name, String description, Int32 sortOrder) {
-			if (board == null) {
-				throw new ArgumentNullException("board");
-			}
-			this.logger.WriteFormat("Create called on CategoryService, Name: {0}, Description: {1}, Sort Order: {2}, board id: {3}", name, description, sortOrder, board.Id);
-			if (!this.IsSolutionAdministrator(this.userProvider.CurrentUser)) {
+		public Category Create(String name, String description, Int32 sortOrder) {
+			this.logger.WriteFormat("Create called on CategoryService, Name: {0}, Description: {1}, Sort Order: {2}", name, description, sortOrder);
+			if (!this.permService.CanCreateCategory(this.userProvider.CurrentUser)) {
 				this.logger.WriteFormat("User does not have permissions to create a new category, name: {0}", name);
-				throw new PermissionException("solution admin");
+				throw new PermissionException("category, create");
 			}
 
 			Category c = new Category {
 				Name = name,
 				SortOrder = sortOrder,
-				Description = description,
-				BoardId = board.Id
+				Description = description
 			};
 
 			this.categoryRepo.Create(c);
@@ -77,7 +67,13 @@ namespace NForum.Core.Services {
 		/// <returns></returns>
 		public Category Read(Int32 id) {
 			this.logger.WriteFormat("Read called on CategoryService, Id: {0}", id);
-			return this.categoryRepo.Read(id);
+			Category category = this.categoryRepo.Read(id);
+			if (!this.permService.HasAccess(this.userProvider.CurrentUser, category, CRUD.Read)) {
+				this.logger.WriteFormat("User does not have permissions to read the category, id: {0}", category.Id);
+				throw new PermissionException("category, read");
+			}
+
+			return category;
 		}
 
 		/// <summary>
@@ -87,20 +83,27 @@ namespace NForum.Core.Services {
 		/// <returns></returns>
 		public Category Read(String name) {
 			this.logger.WriteFormat("Read called on CategoryService, name: {0}", name);
-			return this.categoryRepo.ByName(name);
+			Category category = this.categoryRepo.ByName(name);
+			if (!this.permService.HasAccess(this.userProvider.CurrentUser, category, CRUD.Read)) {
+				this.logger.WriteFormat("User does not have permissions to read the category, name: {0}", category.Name);
+				throw new PermissionException("category, read");
+			}
+
+			return category;
 		}
 
 		/// <summary>
-		/// Method for reading categories by their parent board.
+		/// Method for reading all categories.
 		/// </summary>
-		/// <param name="board">The parent board of the categories to read.</param>
 		/// <returns></returns>
-		public IEnumerable<Category> Read(Board board) {
-			if (board == null) {
-				throw new ArgumentNullException("board");
-			}
-			this.logger.WriteFormat("Read called on CategoryService, board Id: {0}", board.Id);
-			return this.categoryRepo.ReadAll().Where(c => c.BoardId == board.Id).ToList();
+		public IEnumerable<Category> Read() {
+			this.logger.Write("Read called on CategoryService");
+			return this.permService.GetAccessible(
+							this.userProvider.CurrentUser,
+							this.categoryRepo
+								.ReadAll()
+								.ToList()
+						);
 		}
 
 		/// <summary>
@@ -115,45 +118,44 @@ namespace NForum.Core.Services {
 			this.logger.WriteFormat("Update called on CategoryService, Id: {0}", category.Id);
 			// Let's get the category from the data-storage!
 			Category oldCategory = this.Read(category.Id);
+			if (oldCategory == null) {
+				this.logger.WriteFormat("Update category failed, no category with the given id was found, Id: {0}", category.Id);
+				throw new ArgumentException("category does not exist");
+			}
 			Category originalCategory = oldCategory.Clone() as Category;
-			if (!this.IsSolutionAdministrator(this.userProvider.CurrentUser)) {
-				this.logger.WriteFormat("User does not have permissions to update a category, name: {0}", category.Name);
-				throw new PermissionException("solution admin");
+			if (!this.permService.HasAccess(this.userProvider.CurrentUser, oldCategory, CRUD.Update)) {
+				this.logger.WriteFormat("User does not have permissions to update a category, id: {1}, name: {0}", category.Name, category.Id);
+				throw new PermissionException("category, update");
 			}
 
-			if (oldCategory != null) {
-				Boolean changed = false;
-				if (category.Name != oldCategory.Name) {
-					oldCategory.Name = category.Name;
-					changed = true;
-				}
-				if (category.SortOrder != oldCategory.SortOrder) {
-					oldCategory.SortOrder = category.SortOrder;
-					changed = true;
-				}
-				if (category.Description != oldCategory.Description) {
-					oldCategory.Description = category.Description;
-					changed = true;
-				}
-				if (category.CustomProperties != oldCategory.CustomProperties) {
-					oldCategory.CustomProperties = category.CustomProperties;
-					changed = true;
-				}
-
-				if (changed) {
-					oldCategory = this.categoryRepo.Update(oldCategory);
-					this.logger.WriteFormat("Board updated in CategoryService, Id: {0}", category.Id);
-					this.eventPublisher.Publish<CategoryUpdated>(new CategoryUpdated {
-						Category = originalCategory,
-						UpdatedCategory = oldCategory
-					});
-					this.logger.WriteFormat("Update events in CategoryService fired, Id: {0}", category.Id);
-				}
-				return oldCategory;
+			Boolean changed = false;
+			if (category.Name != oldCategory.Name) {
+				oldCategory.Name = category.Name;
+				changed = true;
 			}
-			this.logger.WriteFormat("Update category failed, no category with the given id was found, Id: {0}", category.Id);
-			// TODO:
-			throw new ApplicationException();
+			if (category.SortOrder != oldCategory.SortOrder) {
+				oldCategory.SortOrder = category.SortOrder;
+				changed = true;
+			}
+			if (category.Description != oldCategory.Description) {
+				oldCategory.Description = category.Description;
+				changed = true;
+			}
+			if (category.CustomProperties != oldCategory.CustomProperties) {
+				oldCategory.CustomProperties = category.CustomProperties;
+				changed = true;
+			}
+
+			if (changed) {
+				oldCategory = this.categoryRepo.Update(oldCategory);
+				this.logger.WriteFormat("Board updated in CategoryService, Id: {0}", category.Id);
+				this.eventPublisher.Publish<CategoryUpdated>(new CategoryUpdated {
+					Category = originalCategory,
+					UpdatedCategory = oldCategory
+				});
+				this.logger.WriteFormat("Update events in CategoryService fired, Id: {0}", category.Id);
+			}
+			return oldCategory;
 		}
 
 		/// <summary>
@@ -161,8 +163,15 @@ namespace NForum.Core.Services {
 		/// </summary>
 		/// <param name="category">The category to delete.</param>
 		public void Delete(Category category) {
+			if (category == null) {
+				throw new ArgumentNullException("category");
+			}
 			this.logger.WriteFormat("Delete called on CategoryService, Id: {0}", category.Id);
-			// TODO: Delete forums, topics, posts, access masks etc etc.
+			if (!this.permService.HasAccess(this.userProvider.CurrentUser, category, CRUD.Delete)) {
+				this.logger.WriteFormat("User does not have permissions to delete a category, id: {1}, name: {0}", category.Name, category.Id);
+				throw new PermissionException("category, delete");
+			}
+			// TODO: forums, topics, posts, attachments, etc
 			this.categoryRepo.Delete(category);
 			this.eventPublisher.Publish<CategoryDeleted>(new CategoryDeleted {
 				Category = category
